feat: splitted NativeProvider into multiple files (provider, namespace, node and helpers) to declutter
