+++ crates/provider/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod kubernetes;
 pub mod native;
 pub mod shared;
